requirements.txt
psycopg2-binary
bs4
urllib3
requests
html5lib==1.1
sqlalchemy